server/requirement.txt
Flask
Flask-Cors
Flask-JWT-Extended
Pillow
numpy
tensorflow
requests
pymongo
python-dotenv
bcrypt
Flask-Limiter
itsdangerous
secure-smtplib
cryptography